pyproject.toml
[project]
name = "landbot-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Enrico Graziani",email = "14198729+erikologic@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = ">=0.115.6,<0.116.0"
pydantic-settings = ">=2.7.1,<3.0.0"
pyjwt = { version = ">=2.10.1,<3.0.0", extras = ["crypto"] }
uvicorn = ">=0.34.0,<0.35.0"
slack-sdk = "^3.34.0"
aiohttp = "^3.11.11"
async-cache = "^1.1.1"
structlog = "^24.4.0"
asgi-correlation-id = "^4.3.4"
opentelemetry-instrumentation-fastapi = "^0.50b0"
opentelemetry-sdk = "^1.29.0"
opentelemetry-instrumentation-aiohttp-client = "^0.50b0"

[tool.poetry.group.test.dependencies]
httpx = "^0.28.1"
pytest = "^8.3.4"
pytest-mock = "^3.14.0"
pytest-asyncio = ">=0.25.2,<0.26.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
dev = "poetry run uvicorn app.main:app --reload"
test = "poetry run pytest"
start = "docker-compose up --build"